+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/hooks.ts
@@ -1,5 +1,5 @@
 import { DataCatalogItem, Group, Template } from "@web/extensions/sidebar/core/types";
-import { generateID } from "@web/extensions/sidebar/utils";
+import { generateID, moveItemDown, moveItemUp } from "@web/extensions/sidebar/utils";
 import { useCallback, useEffect, useState } from "react";
 
 import { mergeOverrides } from "../../../hooks";
@@ -142,11 +142,30 @@ export default ({
     onFieldAdd: handleFieldAdd,
   });
 
+  const handleMoveUp = useCallback(
+    (idx: number) => {
+      const newComponents = moveItemUp(idx, dataset.components) ?? dataset.components;
+
+      onDatasetUpdate({ ...dataset, components: newComponents });
+    },
+    [dataset, onDatasetUpdate],
+  );
+
+  const handleMoveDown = useCallback(
+    (idx: number) => {
+      const newComponents = moveItemDown(idx, dataset.components) ?? dataset.components;
+      onDatasetUpdate({ ...dataset, components: newComponents });
+    },
+    [dataset, onDatasetUpdate],
+  );
+
   return {
     activeComponentIDs,
     fieldComponentsList,
     handleFieldUpdate,
     handleFieldRemove,
+    handleMoveUp,
+    handleMoveDown,
     handleCurrentGroupUpdate,
     handleGroupsUpdate,
   };
